using-python/requirements.txt
psycopg2-binary==2.8.5
boto3
